+++ flutter_extensions/lib/flutter_extensions.dart
@@ -1,7 +1,5 @@
 library flutter_extensions;
 
-export 'package:pure_extensions/pure_extensions.dart';
-
 export 'src/animations.dart';
 export 'src/context.dart';
 export 'src/date_time.dart';