+++ pure_extensions/lib/src/dart/collections/iterable.dart
@@ -28,15 +28,6 @@ extension IterablePureExtensions<E> on Iterable<E> {
   /// Concatenates the elements.
   Iterable<E> joinElement(E element) => joinBy((index) => element);
 
-  /// Returns the [index]th element or null if it not exist.
-  E? elementAtOrNull(int index) {
-    try {
-      return elementAt(index);
-    } on RangeError {
-      return null;
-    }
-  }
-
   /// Returns true if the specified value is equal to at least one element of the given list;
   /// false otherwise
   @Deprecated('In favour of [containsEvery]')